pyproject.toml
[project]
name = "odoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "pytx",
    "pytz>=2026.1.post1",
]
